feat: Silence logs by default.
Right now, all log level messages are logged to the output and hence
becomes very noisy. Given that the goal of picoinit is to be a minimal
init system which remains unnoticeable for the most part, this is
against that goal.

Instead silence the logs by default so that only Warning, Error and
Fatal log messages get logged. The pico library also ensures that these
level logs are used only for something critical.

In the future, we might add a flag to picoinit to control this logging
level with the default still remaining the same as Warn.